\name{SegmentationData-package}
\alias{SegmentationData-package}
\alias{SegmentationData}
\docType{package}
\title{
What the package does (short line)
Learning To Talk | Parsing Segmented TextGrids
}
\description{
Parsing and analyzing segmented TextGrids that were 
created as a part of the Learning To Talk project.
}
\details{
\tabular{ll}{
Package: \tab SegmentationData\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-05-17\cr
License: \tab GPL-2\cr
Depends: \tab methods, Filenames, TextGrid\cr
}

}

\author{
Patrick Reidy

Maintainer: Patrick Reidy <patrick.francis.reidy@gmail.com>
}

\references{
\url{learningtotalk.org}
\url{http://www.fon.hum.uva.nl/praat/}
}

\keyword{package}

\seealso{
\code{\linkS4class{TextGrid}}
}

